Final Project & Test Input/Output Files
The src folder contains all the Java code needed to run the program.  The transaction.log and inventory.dat files must be local with the src folder for the program to work; otherwise, the program will prompt you for the name or path to the input files you want to use (or you can change them at the top of the program).  The redbox_kiosk.txt is the final inventory report and the error.log is the error log.  Samples of the output files are included for final output.  Cheers!